Fixed implicit conversion loses integer precision error
Found the error when building the framework Carthage.

Error Message:  implicit conversion loses integer precision: 'int64_t' (aka 'long long') to 'NSInteger' (aka 'int') [-Werror,-Wshorten-64-to-32]

+++ Objective-C/CBLQuery.mm
@@ -350,7 +350,7 @@ C4LogDomain QueryLog;
 }
 
 - (NSInteger) integerAtIndex: (NSUInteger)index {
-    return FLValue_AsInt(FLArray_Get(_customColumns, (uint32_t)index));
+    return (NSInteger)FLValue_AsInt(FLArray_Get(_customColumns, (uint32_t)index));
 }
 
 - (float) floatAtIndex: (NSUInteger)index {
